+++ src/constants.rs
@@ -1,0 +1,17 @@
+pub const BOOT_PARTITION_INDEX: u8 = 1;
+pub const ROOT_PARTITION_INDEX: u8 = 3;
+
+pub static JOURNALD_CONF: &str = "
+[Journal]
+Storage=volatile
+SystemMaxUse=16M
+";
+
+pub const BASE_PACKAGES: [&str; 6] = [
+    "base",
+    "grub",
+    "efibootmgr",
+    "intel-ucode",
+    "networkmanager",
+    "broadcom-wl",
+];
