+++ src/storage/markers.rs
@@ -1,5 +1,5 @@
 use std::path::Path;
-
+// Marker traits
 pub trait BlockDevice: std::fmt::Debug {
     fn path(&self) -> &Path;
 }
