+++ src/tool/mod.rs
@@ -1,4 +1,12 @@
-use super::error::*;
+mod chroot;
+mod mount;
+mod qemu;
+
+pub use chroot::chroot;
+pub use mount::mount;
+pub use qemu::qemu;
+
+use crate::error::*;
 use failure::ResultExt;
 use std::path::PathBuf;
 use std::process::Command;
